BoardService.java - 수정,삭제 유효성 추가

+++ src/main/java/com/example/myweb/service/BoardService.java
@@ -1,5 +1,6 @@
 package com.example.myweb.service;
 
+import com.example.myweb.config.auth.PrincipalDetail;
 import com.example.myweb.dto.ReplyDto;
 import com.example.myweb.model.Board;
 import com.example.myweb.model.Reply;
@@ -8,13 +9,14 @@ import com.example.myweb.model.User;
 import com.example.myweb.repository.BoardRepository;
 import com.example.myweb.repository.ReplyRepository;
 import com.example.myweb.repository.ReplyToCommentRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-
+@Slf4j
 @Service
 public class BoardService {
 
@@ -46,15 +48,25 @@ public class BoardService {
     }
 
     @Transactional
-    public void boardDelete(Long id) {
+    public void boardDelete(Long id, User user) {
+        Board board = boardRepository.findById(id).orElseThrow(()->{
+            return new IllegalArgumentException("글 찾기 실패 : 아이디를 찾을 수 없습니다.");
+        });
+
+        if(!(board.getUser().equals(user))) {
+            throw new IllegalArgumentException("인증 실패 : 삭제할 수 없습니다.");
+        }
         boardRepository.deleteById(id);
     }
 
     @Transactional
-    public void boardUpdate(Long id, Board requestBoard) {
+    public void boardUpdate(Long id, Board requestBoard, User user) {
         Board board = boardRepository.findById(id).orElseThrow(()->{
             return new IllegalArgumentException("글 찾기 실패 : 아이디를 찾을 수 없습니다.");
         });
+        if(!(board.getUser().equals(user))) {
+            throw new IllegalArgumentException("인증 실패 : 수정할 수 없습니다.");
+        }
         board.setTitle(requestBoard.getTitle());
         board.setContent(requestBoard.getContent());
     }
@@ -62,7 +74,7 @@ public class BoardService {
     @Transactional
     public void replySave(Long boardId, ReplyDto requestReply, User user) {
         Board board = boardRepository.findById(boardId).orElseThrow(()->{
-            return new IllegalArgumentException("게시글을 찾을 수 없습니다.");
+            return new IllegalArgumentException("글 찾기 실패 : 아이디를 찾을 수 없습니다.");
         });
 
         Reply reply = new Reply();
@@ -73,7 +85,13 @@ public class BoardService {
         replyRepository.save(reply);
     }
 
-    public void replyDelete(Long replyId) {
+    public void replyDelete(Long replyId, User user) {
+        Reply reply = replyRepository.findById(replyId).orElseThrow(()->{
+            return new IllegalArgumentException("댓글 찾기 실패 : 아이디를 찾을 수 없습니다.");
+        });
+        if(!(reply.getUser().equals(user))) {
+            throw new IllegalArgumentException("인증 실패 : 수정할 수 없습니다.");
+        }
         replyRepository.deleteById(replyId);
     }
 
@@ -91,7 +109,13 @@ public class BoardService {
         replyToCommentRepository.save(RTC);
     }
 
-    public void replyToCommentDelete(Long replyToCommentId) {
-        replyRepository.deleteById(replyToCommentId);
+    public void replyToCommentDelete(Long replyToCommentId, User user) {
+        ReplyToComment reply = replyToCommentRepository.findById(replyToCommentId).orElseThrow(()->{
+            return new IllegalArgumentException("댓글을 찾을 수 없습니다.");
+        });
+        if(!(reply.getUser().equals(user))) {
+            throw new IllegalArgumentException("인증 실패 : 삭제할 수 없습니다.");
+        }
+        replyToCommentRepository.deleteById(replyToCommentId);
     }
 }
